% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmdbConnection.R
\name{connect,mnmdb::mnmdbConnection-method}
\alias{connect,mnmdb::mnmdbConnection-method}
\title{Connect a mnmdbConnection object to the actual database using \code{auth}.}
\usage{
## S7 method for class <mnmdb::mnmdbConnection>
connect(conn)
}
\arguments{
\item{conn}{a mnmdbConnection connection to be connected}
}
\value{
mnmdbConnection
}
\description{
Connect a mnmdbConnection object to the actual database using \code{auth}.
}
\details{
This function can be used in a pipe:
it receives and returns a mnmdbConnection.
}
