% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmdbAuth.R
\name{describe,mnmdb::mnmdbAuth-method}
\alias{describe,mnmdb::mnmdbAuth-method}
\title{description of database auth}
\usage{
## S7 method for class <mnmdb::mnmdbAuth>
describe(x)
}
\description{
description of database auth
}
