% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmdbConnection.R
\name{describe,mnmdb::mnmdbConnection-method}
\alias{describe,mnmdb::mnmdbConnection-method}
\title{description of database connection}
\usage{
## S7 method for class <mnmdb::mnmdbConnection>
describe(x)
}
\description{
description of database connection
}
