% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R, R/mnmdbAuth.R, R/mnmdbConnection.R
\name{describe}
\alias{describe}
\alias{describe,mnmdb::mnmdbAuth-method}
\alias{describe,mnmdb::mnmdbConnection-method}
\title{generic for description of mnmdb objects}
\usage{
describe(x, ...)

## S7 method for class <mnmdb::mnmdbAuth>
describe(x)

## S7 method for class <mnmdb::mnmdbConnection>
describe(x)
}
\arguments{
\item{x}{the connection to be described}

\item{...}{Not used.}
}
\value{
descriptive string
}
\description{
Some text describing the content of an mnmdb object
}
