% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring_management.R
\name{init_keyring}
\alias{init_keyring}
\title{Initialize keyring to store mnmdb credentials.}
\usage{
init_keyring(keyring_label = "mnmdb_temp")
}
\arguments{
\item{keyring_label}{label of the keyring to lock; default is "mnmdb_temp"}
}
\description{
This function will initialize a keyring (unless it already exists)
and unlock it for temporary use.
}
