% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table.R
\name{query_table}
\alias{query_table}
\alias{query_table,mnmdb::mnmdbConnection-method}
\title{Query tables from MNE databases.}
\usage{
query_table(conn, ...)

## S7 method for class <mnmdb::mnmdbConnection>
query_table(conn, table_id, subselect = NA)
}
\arguments{
\item{conn}{the mnmdb connection}

\item{...}{Not used.}

\item{table_id}{the \code{DBI::Id} of a table}

\item{subselect}{character vector of fields to select}
}
\value{
tibble or (spatial) data frame
}
\description{
This function enables direct query of data from MNE databases.
The \code{subselect} arg can be used to query only a subset of fields.
}
