% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring_management.R
\name{terminate_keyring}
\alias{terminate_keyring}
\title{terminate and clean up a given keyring}
\usage{
terminate_keyring(keyring_label = "mnmdb_temp")
}
\arguments{
\item{keyring}{keyring label to empty (do not use \code{Login}!)}
}
\description{
This function will loop through all labels in a given keyring,
deleting all keys it encounters.
The purpose is to leave no passwords in store after a session ends.
Tip: the \code{seahorse} util can be used to browse the system keyring.
}
