% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring_management.R
\name{unlock_keyring}
\alias{unlock_keyring}
\title{Unlock a keyring, and schedule locking for when the R session ends.}
\usage{
unlock_keyring(keyring_label = "mnmdb_temp", ...)
}
\arguments{
\item{...}{additional parameters for \code{keyring::keyring_unlock()}}

\item{keyring}{label of the keyring to lock}
}
\description{
This function will unlock a given keyring, but immediately schedule
locking for later.
}
