% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring_management.R
\name{lock_keyring_delayed}
\alias{lock_keyring_delayed}
\title{Lock the keyring after a delay. (only on Linux)}
\usage{
lock_keyring_delayed(keyring_label = "mnmdb_temp", delay = 3600)
}
\arguments{
\item{keyring_label}{label of the keyring to lock}

\item{delay}{wait time (in seconds)}
}
\description{
This function will launch a background process via \code{system}
to wait for a given time before locking a keyring.
}
