% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmdbAuth.R
\name{mnmdbAuth}
\alias{mnmdbAuth}
\title{Authentication for mnmdb via connection parameters}
\usage{
mnmdbAuth(
  folder = character(0),
  host = "127.0.0.1",
  port = "5432",
  database = character(0),
  user = character(0),
  password = character(0)
)
}
\arguments{
\item{folder}{the structure folder with csv files for all tables}

\item{host}{the host (IP address or url); default 'localhost'}

\item{port}{on which the host listens for input; default '5439'}

\item{database}{name of the database to access}

\item{user}{username to get connected}
}
\description{
manage authentification storage for connections to mnm databases
}
\details{
tbd
}
\examples{
\dontrun{
  auth <- mnmdb::mnmdbAuth(file.path("config_files", "test.conf"))
}

}
