% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmdbAuth.R
\name{mnmdbAuth}
\alias{mnmdbAuth}
\title{Authentication for mnmdb via connection parameters}
\usage{
mnmdbAuth(...)
}
\arguments{
\item{...}{not used (consturctor)}

\item{host}{the host address (IP or url); default '127.0.0.1'}

\item{port}{on which the host listens for input; default '5432'}

\item{database}{name of the database to access (mandatory)}

\item{user}{username to get connected (mandatory)}

\item{password}{(optional) password to connect;
will be prompted if no password is provided by the user}

\item{connect_passwordless}{option to connect without password via .pgpass}
}
\description{
Manage authentification storage for connections to mnm databases.
In particular, the password is stored in the system keyring, and
never leaves the constructor function.
}
\examples{
\dontrun{
  auth <- mnmdb::mnmdbAuth(
    user = "guest",
    database = "sandbox",
    connect_passwordless = TRUE # optional for use of `~/.pgpass`
  )
  auth <- mnmdbAuth(
    host = "127.0.0.1",
    port = "5432",
    user = "guest",
    password = "abc123", # avoid and do not share plaintext passwords!
    database = "sandbox"
  )
}

}
