% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmdbAuth.R
\name{mnmdbAuth}
\alias{mnmdbAuth}
\title{Authentication for mnmdb via config file}
\usage{
mnmdbAuth(config_file = character(0))
}
\arguments{
\item{config_file}{file.path to a config file to load with \code{configr}}
}
\description{
manage authentification storage for connections to mnm databases
}
\details{
tbd
}
