% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmdbAuth.R
\name{mnmdbAuthConf}
\alias{mnmdbAuthConf}
\title{Authentication for mnmdb via config file}
\usage{
mnmdbAuthConf(
  folder = character(0),
  host = character(0),
  port = character(0),
  database = character(0),
  user = character(0),
  config_file = character(0)
)
}
\arguments{
\item{folder}{the structure folder with csv files for all tables}

\item{host}{the host (IP address or url); default 'localhost'}

\item{port}{on which the host listens for input; default '5439'}

\item{database}{name of the database to access}

\item{user}{username to get connected}

\item{config_file}{file.path to a config file to load with \code{configr}}
}
\description{
Authentication for mnmdb via config file
}
\examples{
\dontrun{
  auth <- mnmdb::mnmdbAuth(file.path("config_files", "test.conf"))
}

}
