% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmdbAuth.R
\name{mnmdbAuthConf}
\alias{mnmdbAuthConf}
\title{Authentication for mnmdb via config file}
\usage{
mnmdbAuthConf(config_file)
}
\arguments{
\item{config_file}{file.path to a config file to load with \code{configr}}
}
\description{
Authentication for mnmdb via config file
}
\examples{
\dontrun{
  auth <- mnmdb::mnmdbAuth(file.path("config_files", "test.conf"))
}

}
