% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmdbAuth.R
\name{mnmdbAuthConf}
\alias{mnmdbAuthConf}
\title{Authentication for mnmdb via config file}
\usage{
mnmdbAuthConf(config_file, ...)
}
\arguments{
\item{config_file}{file.path to a config file to load with \code{configr}}

\item{...}{not used (consturctor)}
}
\description{
Just as \code{mnmdbAuth}, but with the extra option to store authentication parameters
in a config file (e.g. in the \code{config_files} folder).
}
\examples{
\dontrun{
  auth <- mnmdb::mnmdbAuth(file.path("config_files", "test.conf"))
}

}
