% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmdbConnection.R
\name{mnmdbConnection}
\alias{mnmdbConnection}
\title{MNE Database Connection Object.}
\usage{
mnmdbConnection(auth = NULL, database_connection = NULL, is_connected = FALSE)
}
\arguments{
\item{auth}{an mnmdbAuth object which stores authentication}
}
\description{
Given an authentication for login, this database object is supposed
to facilitate all interactions with an actual database.
}
\examples{
\dontrun{
  test_auth <- mnmdbAuth(
    user = "guest",
    database = "sandbox"
  )
  mnmdbconn <- mnmdbConnection(auth <- test_auth)
  mnmdbconn <- mnmdbconn |> connect()
  # dplyr::tbl(mnmdbconn@database_connection, DBI::Id("test", "mtcars"))
}

}
