% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_pkgs.R
\name{require_pkgs}
\alias{require_pkgs}
\title{Check availability of required packages}
\usage{
require_pkgs(pkgs, ...)
}
\arguments{
\item{pkgs}{A character vector of package names.}

\item{...}{more parameters for base::requireNamespace
(e.g. \code{quietly} for non-verbose check for requirements)}
}
\description{
Takes a vector of package names and passes each name to
\code{\link[base:ns-load]{requireNamespace()}};
if package(s) are missing, returns an error message providing the basic
\code{install.packages()} command to install them.
}
\examples{
\dontrun{
  require_pkgs(c("a", "base", "b", "magrittr"))
  # not_attached <- any(devtools::loaded_packages() == "magrittr") == FALSE
}

}
\keyword{internal}
