% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring_management.R
\name{store_db_password}
\alias{store_db_password}
\title{Set a password in a keyring, prompt one if none is provided.}
\usage{
store_db_password(keyring_label = "mnmdb_temp", ..., password = NULL)
}
\arguments{
\item{password}{password (NULL to query one)}

\item{keyring}{label of the keyring to lock}

\item{service}{"service" identifier of the keyring}

\item{username}{"username" who holds the key}
}
\description{
Prompt the user for a password (if none was given) and
store the password to the mnmdb credential keyring
with function arguments to \code{keyring::key_set_with_value}
}
